[Refactor]: 캘린더용 DTO 패키지 분리
