+++ src/main/java/scs/planus/domain/todo/service/GroupTodoCalendarService.java
@@ -7,10 +7,10 @@ import org.springframework.transaction.annotation.Transactional;
 import scs.planus.domain.group.entity.GroupMember;
 import scs.planus.domain.group.repository.GroupMemberRepository;
 import scs.planus.domain.group.repository.GroupRepository;
-import scs.planus.domain.todo.dto.TodoDailyDto;
-import scs.planus.domain.todo.dto.TodoDailyResponseDto;
-import scs.planus.domain.todo.dto.TodoDailyScheduleDto;
-import scs.planus.domain.todo.dto.TodoPeriodResponseDto;
+import scs.planus.domain.todo.dto.calendar.TodoDailyDto;
+import scs.planus.domain.todo.dto.calendar.TodoDailyResponseDto;
+import scs.planus.domain.todo.dto.calendar.TodoDailyScheduleDto;
+import scs.planus.domain.todo.dto.calendar.TodoPeriodResponseDto;
 import scs.planus.domain.todo.entity.GroupTodo;
 import scs.planus.domain.todo.repository.TodoQueryRepository;
 import scs.planus.global.exception.PlanusException;
